DetailWeatherViewController refactored
